backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Ryan Lin  <linjinhong@yandex.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.1.0"
flask-cors = "^5.0.1"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.1"
supabase = "^2.15.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
